+++ odps-console-dship/src/main/java/com/aliyun/odps/ship/upload/BlockRecordReader.java
@@ -32,22 +32,16 @@ import org.apache.commons.io.IOUtils;
 import com.aliyun.odps.ship.common.BlockInfo;
 import com.aliyun.odps.ship.common.Constants;
 
-public class BlockRecordReader {
+public class BlockRecordReader extends RecordReader {
 
-  private BlockInfo blockInfo;
   private byte[] fieldDelimiter;
   private byte[] recordDelimiter;
   boolean ignoreHeader;
+  boolean isLastLine;
+  byte [] currentLine;
 
 
   private BufferedInputStream is;
-  private long startPos;
-  private long readBytes;
-  private byte[] currentLine;
-  boolean isLastLine;
-
-  String detectedCharset;
-  int bomBytes;
 
   private ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream(8 * 1024 * 1024);
   private final int BUF_SIZE = 8 * 1024 * 1024;
@@ -57,7 +51,8 @@ public class BlockRecordReader {
 
   public BlockRecordReader(BlockInfo blockInfo, String fd, String rd, boolean ignoreHeader)
       throws IOException {
-    this.blockInfo = blockInfo;
+    super(blockInfo);
+
     this.fieldDelimiter = fd.getBytes();
     this.recordDelimiter = rd.getBytes();
     this.ignoreHeader = ignoreHeader;
@@ -65,7 +60,7 @@ public class BlockRecordReader {
     init();
   }
 
-  public byte[][] readTextRecord() throws IOException, ParseException {
+  public byte[][] readTextRecord() throws IOException {
     if (isLastLine) {
       return null;
     }
@@ -117,7 +112,7 @@ public class BlockRecordReader {
           lineLength += bufLength - offset - recordDelimiter.length;
           if (lineLength > Constants.MAX_RECORD_SIZE) {
             throw new IllegalArgumentException(
-                Constants.ERROR_INDICATOR + "line big than 200M - please check record delimiter");
+                Constants.ERROR_INDICATOR + "line bigger than 200M - please check record delimiter");
           }
           byteArrayOutputStream.write(buf, offset, bufLength - recordDelimiter.length - offset);
           offset = bufLength - recordDelimiter.length;
@@ -152,16 +147,9 @@ public class BlockRecordReader {
     }
   }
 
-  public long getReadBytes() {
-    return readBytes;
-  }
 
   public void close() throws IOException {
     is.close();
-  }
-
-  public String getDetectedCharset() {
-    return detectedCharset;
   }
 
   public static int indexOf(byte[] src, int offset, int length, byte[] search) {
@@ -216,49 +204,5 @@ public class BlockRecordReader {
       readLine();
     }
     isLastLine = false;
-  }
-
-  /**
-   *
-   * http://www.unicode.org/unicode/faq/utf_bom.html BOMs: 00 00 FE FF = UTF-32, big-endian FF FE 00
-   * 00 = UTF-32, little-endian EF BB BF = UTF-8, FE FF = UTF-16, big-endian FF FE = UTF-16,
-   * little-endian
-   *
-   ***/
-  /**
-   * Read four bytes and check for BOM marks.
-   */
-  private void detectBomCharset() throws IOException {
-    InputStream internalIs = blockInfo.getFileInputStream();
-    try {
-      byte bom[] = new byte[4];
-      int n = internalIs.read(bom, 0, bom.length);
-
-      if ((n >= 4) && (bom[0] == (byte) 0x00) && (bom[1] == (byte) 0x00) &&
-          (bom[2] == (byte) 0xFE) && (bom[3] == (byte) 0xFF)) {
-        detectedCharset = "UTF-32BE";
-        bomBytes = 4;
-      } else if ((n >= 4) && (bom[0] == (byte) 0xFF) && (bom[1] == (byte) 0xFE) &&
-                 (bom[2] == (byte) 0x00) && (bom[3] == (byte) 0x00)) {
-        detectedCharset = "UTF-32LE";
-        bomBytes = 4;
-      } else if ((n >= 3) && (bom[0] == (byte) 0xEF) && (bom[1] == (byte) 0xBB) &&
-                 (bom[2] == (byte) 0xBF)) {
-        detectedCharset = "UTF-8";
-        bomBytes = 3;
-      } else if ((n >= 2) && (bom[0] == (byte) 0xFE) && (bom[1] == (byte) 0xFF)) {
-        detectedCharset = "UTF-16BE";
-        bomBytes = 2;
-      } else if ((n >= 2) && (bom[0] == (byte) 0xFF) && (bom[1] == (byte) 0xFE)) {
-        detectedCharset = "UTF-16LE";
-        bomBytes = 2;
-      } else {
-        // Unicode BOM mark not found, unread all bytes
-        detectedCharset = null;
-        bomBytes = 0;
-      }
-    } finally {
-      IOUtils.closeQuietly(internalIs);
-    }
   }
 }